[back/front] Deletion of an external reference linked to a file --> deletion of the file (#2417)
