+++ opencti-platform/opencti-front/src/private/components/analysis/ExternalReferences.tsx
@@ -5,12 +5,13 @@ import { saveViewParameters } from '../../../utils/ListParameters';
 import ListLines from '../../../components/list_lines/ListLines';
 import ExternalReferencesLines, { externalReferencesLinesQuery } from './external_references/ExternalReferencesLines';
 import ExternalReferenceCreation from './external_references/ExternalReferenceCreation';
-import Security, { KNOWLEDGE_KNUPDATE } from '../../../utils/Security';
+import Security from '../../../utils/Security';
 import useLocalStorage, { localStorageToPaginationOptions } from '../../../utils/hooks/useLocalStorage';
 import {
   ExternalReferencesLinesPaginationQuery$data,
   ExternalReferencesLinesPaginationQuery$variables,
 } from './external_references/__generated__/ExternalReferencesLinesPaginationQuery.graphql';
+import { KNOWLEDGE_KNUPDATE } from '../../../utils/hooks/useGranted';
 
 export const externalReferencesSearchQuery = graphql`
   query ExternalReferencesSearchQuery(
