+++ opencti-platform/opencti-front/src/private/components/analysis/external_references/ExternalReferenceCreation.tsx
@@ -91,6 +91,7 @@ export const externalReferenceCreationMutation = graphql`
       url
       external_id
       created
+      fileId
     }
   }
 `;
