--- opencti-platform/opencti-front/src/private/components/analysis/external_references/ExternalReferenceHeader.js
@@ -1,90 +1,0 @@
-import React, { Component } from 'react';
-import * as PropTypes from 'prop-types';
-import { compose } from 'ramda';
-import { graphql, createFragmentContainer } from 'react-relay';
-import withStyles from '@mui/styles/withStyles';
-import Typography from '@mui/material/Typography';
-import { truncate } from '../../../../utils/String';
-import inject18n from '../../../../components/i18n';
-import Security from '../../../../utils/Security';
-import { KNOWLEDGE_KNUPDATE } from '../../../../utils/hooks/useGranted';
-
-const styles = () => ({
-  title: {
-    float: 'left',
-  },
-  popover: {
-    float: 'left',
-    marginTop: '-13px',
-  },
-  aliases: {
-    marginRight: 7,
-  },
-  aliasesInput: {
-    margin: '4px 0 0 10px',
-    float: 'right',
-  },
-  modes: {
-    margin: '-10px 0 0 0',
-    float: 'right',
-  },
-  button: {
-    marginRight: 20,
-  },
-  export: {
-    margin: '-10px 0 0 0',
-    float: 'right',
-  },
-});
-
-class ExternalReferenceHeaderComponent extends Component {
-  render() {
-    const { classes, externalReference, PopoverComponent } = this.props;
-    return (
-      <div>
-        <Typography
-          variant="h1"
-          gutterBottom={true}
-          classes={{ root: classes.title }}
-        >
-          {truncate(externalReference.source_name, 80)}
-        </Typography>
-        <Security needs={[KNOWLEDGE_KNUPDATE]}>
-          <div className={classes.popover}>
-            {React.cloneElement(PopoverComponent, { id: externalReference.id })}
-          </div>
-        </Security>
-        <div className="clearfix" />
-      </div>
-    );
-  }
-}
-
-ExternalReferenceHeaderComponent.propTypes = {
-  externalReference: PropTypes.object,
-  PopoverComponent: PropTypes.object,
-  variant: PropTypes.string,
-  classes: PropTypes.object,
-  t: PropTypes.func,
-  fld: PropTypes.func,
-  link: PropTypes.string,
-  modes: PropTypes.array,
-  currentMode: PropTypes.string,
-  knowledge: PropTypes.bool,
-  adjust: PropTypes.func,
-};
-
-const ExternalReferenceHeader = createFragmentContainer(
-  ExternalReferenceHeaderComponent,
-  {
-    externalReference: graphql`
-      fragment ExternalReferenceHeader_externalReference on ExternalReference {
-        id
-        source_name
-        description
-      }
-    `,
-  },
-);
-
-export default compose(inject18n, withStyles(styles))(ExternalReferenceHeader);
